Improved Table Summaries, Removed QuadWord Class
Summaries now detail every table entry for a given Hat, and moved to uint arrays from quadwords since it's just simpler overall.
Additionally, reworked the logic for determining whether or not a InfoPack is default or not, and added the Alloys and new/in-dev EX characters to the internal names lists.