HatManager Discards REL and Arr After Population
Previously the Hat Manager would load the Kirby REL itself, dump the tables from it, and keep bot the RELNode and table byte array in memory; this is no longer the case. Instead, we parse a single time to populate the dictionary of entries, then discard both.

On export, we then construct the outgoing table file based on the state of the entry dictionary at the time the export is requested.

Also split the REL load function into two overloads, one which accepts a filepath and grabs the REL itself, and another which directly accepts a REL node.