Fixed Hat Cloning Issue, Cleaned Up Hat Functions
Previously I was incorrectly using the assignment operator from one hat slot into another slot when cloning or creating new hats; *this is not okay in C#* lmfao. In C# that makes a second reference to the same hat, which meant that clones essentially shared data before! This is fixed, shouldn't be a problem anymore.

Also, deleted the old "copyHatToEmptySlot" function in favor of just adding a "allowOverwrite" parameter to the main copyHatToSlot function; serves the same purpose essentially.