Impl. Loading Names from FighterConfig Dir
Implements functionality which detects whether or not an incoming .rel file is located within a "/module/" folder, and if so, attempts to access the neighboring "/BrawlEx/FighterConfig/" directory to get the names associated with different Fighter IDs from there!

Additionally, set up default settings so that the program will first attempt to load a Kirby module from its own folder, and failing that will instead attempt to load from the "pf/module/" folder within a build; invoking the above functionality for getting Fighter Names from the Config folder if it succeeds!