Standardized TXT and XML Importer Interfaces
Was looking at the two classes and liked the option to parse but not import offered by the TXT class, as well as the ability to directly import in the XML class, so mirrored the respective functionalities into each class. Also matched up some of the wording between the two as well, just for consistency.

This also entailed changing the way the TXT import function in Form1 actually works, since we now use the direct import option for that. Also, can now import from multiple files at the same time, for both file types.

Lastly, minor adjustment to the way setting names works in HatNames: if you try to set a name to a null or empty string, it'll instead remove the name for the target ID.