Implemented PropGetter Class
After the separation of names from the HatNode class, the old approach of just popping the HatInfoPack into the PropertyGrid itself lost the ability to provide an interface for editing a Hat's name. To remedy this while keeping things unified in the property grid itself, a new class has been implemented which just functions as a thing for the PropertyGrid to attach to and read everything it needs from.