XML Export No Longer Hardcodes Source Dictionaries
Previously the export function explicitly took in the hat dictionary from the passed in HatManager and the name dictionary from the HatNames struct, which was kinda limiting to how you could export things. In particular it woulda required any potential applications that just want the XML export import functionality to also include the HatNames struct, which I didn't like either.

Only kinda ugly thing about this is that it's required making the HatNames internal dictionary public, which I don't like but whatever lol.