Force-Partial Check Ignores Tertiary Version Field
Reworked the check for forcing XML Hats Imports to only care about the Primary and Secondary version fields, ignoring the Tertiary field (eg. the 0 in 1.1.*0*). This way we can still use that field to signify things like minor bugfix updates which don't introduce changes to the REL layout or anything without triggering a potentially undesirably forcing on Partial mode.

As part of this, I've made KirbyHatManager.cs the thing which dictates the current version number so that the logic for forcing Partial mode is hopefully portable.