XML Export now Emits Empty Nodes for Null WriteCmd
Previously, the XML Exporter didn't emit anything at all for empty WriteCmd entries; it'd just check if they were invalid and skip if so. This causes some mild issues with the partial hat concept, since that process interprets a non-existent node as something to preserve from the existing slot.

For the most part this isn't actually a problem, as for the most part people are either going to be importing custom hats over empty slots or existing characters over the slots they already occupy. However, in situations for example where people are importing exported hats over a REL which contains optional characters (like the base Ex REL), that *could* result in situations where characters mistakenly inherit WriteCmds where they shouldn't be; so to address that we now explicitly emit empty nodes instead of nothing.