Set Up Proper Hat Version Promoter Logic
Thought deeper about the partial hat system as a solution to the 1.0.1 -> 1.1.0 migration process, and I think the cracks in that approach just aren't ultimately acceptable. Looking over a real-life example of a Hat XML that'd need to be migratable, there was an example of a Hat that would have broken in the process as well as several others that could have broken if the Hats they imported over were configured differently. That last one being particularly nasty, since it would have meant that migration compatibility would have been dependent on the state of the new base REL, which just isn't predictable or acceptable.
This should fix all that; the new system explicitly notes version where layout changes have happened, and specifies what fields need to be inherited when promoting up through that version. Should also be easily extensible, so hopefully we won't need to rework this whole system next time a table gets added.
Also got rid of the global forcePartial functionality, might bring it back if it turns out it would be meaningfully useful in the future.